Initial Layout Commit
In this commit, I did the initial (and probably main) grid-work,
defining every div with its grid inside, and then giving html code
structure to the navigation bar as well as the menu (that stays in the
left side).

The next commit should have the main grid organized with its inner grids
and elements, with the small navigation bar.