A lot of changes...
But i just noticed that the youtube layout actualy has only 2 columns
(even though the project only has two)...
and that the menu on the top left as well as the logo belong to the
first column (? maybe, still figuring out).

Added the javaScript.js file and some code related to the search bar.
Now it shows an icon on the left-side, cool? not the best one but
definitily something ;D

Now I need to check what to do with my grids.